+++ neo/aas/AASFile.cpp
@@ -1488,9 +1488,9 @@ bool idAASFileLocal::Load( const idStr& fileName, unsigned int mapFileCRC )
 idAASFileLocal::MemorySize
 ================
 */
-int idAASFileLocal::MemorySize() const
+size_t idAASFileLocal::MemorySize() const
 {
-	int size;
+	size_t size;
 	
 	size = planeList.Size();
 	size += vertices.Size();
